Add files via upload
A function in Java is a block of code that performs a specific task and can be reused in different parts of the program. A function is also known as a method in Java, and it must be defined within a class. A function can take one or more parameters, which are values passed to the function when it is called, and it can return a value, which is the result of the function’s computation.